+++ src/gameObject/fruits/Banana.java
@@ -35,6 +35,6 @@ public class Banana extends Fruit {
 		{
 		    System.out.println("Banana rendering failed");
 		}
-		graphics.drawImage(img1, xCoordinate,yCoordinate, null);
+		graphics.drawImage(img1, (int)xCoordinate,(int)yCoordinate, null);
     }
 }
