+++ src/gameObject/fruits/Peach.java
@@ -34,6 +34,6 @@ public class Peach extends Fruit {
 		{
 		    System.out.println("Peach rendering failed");
 		}
-		graphics.drawImage(img1, xCoordinate,yCoordinate, null);
+		graphics.drawImage(img1, (int)xCoordinate,(int)yCoordinate, null);
     }
 }
